+++ test/fixtures/runtime-params-invalid-hook/src/pages/index.tsx
@@ -1,4 +1,4 @@
-import { useParams } from "@kudzujs/core"
+import { useParams } from "react-router-dom"
 
 export default function Page() {
   const params = useParams()